# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Batuan, Fosil dan Sumber Daya Geologi",
        "title": "Museum Geologi Bandung",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://sbc-sebatcabut.herokuapp.com/",
    "basePath": "/v1",
    "paths": {
        "/batuan": {
            "post": {
                "description": "Create a new Batuan with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batuan"
                ],
                "summary": "Create a new Batuan",
                "parameters": [
                    {
                        "description": "The batuan to create",
                        "name": "batuan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Batuan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    }
                }
            }
        },
        "/batuan/{batuanId}": {
            "get": {
                "description": "Get a Batuan by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batuan"
                ],
                "summary": "Get Batuan by ID",
                "operationId": "get-batuan-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batuan ID",
                        "name": "batuanId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit an existing Batuan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batuan"
                ],
                "summary": "Edit an existing Batuan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the Batuan to edit",
                        "name": "batuanId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Batuan object that needs to be edited",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Batuan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a batuan by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batuan"
                ],
                "summary": "Delete a batuan by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the batuan to delete",
                        "name": "batuanId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    }
                }
            }
        },
        "/batuans": {
            "get": {
                "description": "Retrieve all batuans from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batuan"
                ],
                "summary": "Get all batuans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    }
                }
            }
        },
        "/batuans/export": {
            "get": {
                "description": "Get data batuan from MongoDB and export to excel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batuan"
                ],
                "summary": "Export data batuan to excel",
                "responses": {
                    "200": {
                        "description": "Data batuan exported to excel successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.BatuanResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.BatuanResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}